pyproject.toml
[tool.poetry]
package-mode = false
name = "who-am-i"
version = "0.1.0"
description = "프로젝트 설명"
authors = ["Creeper Eye <iostream1204@gmail.com>"]
license = "creepereye"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.1"
djangorestframework = "^3.15.2"
uwsgi = "^2.0.27"
psycopg2-binary = "^2.9.9"
pillow = "^11.0.0"
django-prometheus = "^2.3.1"


[tool.poetry.group.dev.dependencies]


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
